Solution to Calculate Employee Bonus

+++ Calculate Employee Bonus@@67877d61dc107486ad5c274f/code.c
@@ -7,7 +7,7 @@ struct data {
 };
 
 char res(float salary){
-    salary<=5000? ((0.1*salary)+salary):((0.05*salary)+salary);
+    salary<=5000? float((0.1*salary)+salary):float((0.05*salary)+salary);
 }
 
 int main() {
